Add files via upload
I learned how to collect dataset from Kaggle and other sources, clean, merge and analyze it—also learned how to make different data visualizations using Tableau and create a story out of it.

We have a common misconception that alcohol, rain, and snow are the main factors contributing to accidents but we learned from the visualization that it is not the case. Many accidents have occurred in clear and cloudy weather conditions and without any alcohol involvement. There are many other factors that contribute to the same. We learned and experienced how visualization can speak for itself.